refactor: renamed subpackage factory to plugin
