+++ sparkdq/checks/aggregate/count_checks/count_min_check.py
@@ -6,7 +6,7 @@ from sparkdq.core.base_config import BaseAggregateCheckConfig
 from sparkdq.core.check_results import AggregateEvaluationResult
 from sparkdq.core.severity import Severity
 from sparkdq.exceptions import InvalidCheckConfigurationError
-from sparkdq.factory.check_config_registry import register_check_config
+from sparkdq.plugin.check_config_registry import register_check_config
 
 
 class RowCountMinCheck(BaseAggregateCheck):
