+++ sparkdq/checks/row_level/contained_checks/is_contained_in_check.py
@@ -6,7 +6,7 @@ from sparkdq.core.base_check import BaseRowCheck
 from sparkdq.core.base_config import BaseRowCheckConfig
 from sparkdq.core.severity import Severity
 from sparkdq.exceptions import InvalidCheckConfigurationError
-from sparkdq.factory.check_config_registry import register_check_config
+from sparkdq.plugin.check_config_registry import register_check_config
 
 
 class IsContainedInCheck(BaseRowCheck):
