+++ sparkdq/checks/row_level/date_checks/date_between_check.py
@@ -7,7 +7,7 @@ from sparkdq.checks.base_classes.base_comparison_check import BaseComparisonChec
 from sparkdq.core.base_config import BaseRowCheckConfig
 from sparkdq.core.severity import Severity
 from sparkdq.exceptions import InvalidCheckConfigurationError
-from sparkdq.factory.check_config_registry import register_check_config
+from sparkdq.plugin.check_config_registry import register_check_config
 
 
 class DateBetweenCheck(BaseComparisonCheck):
