+++ sparkdq/checks/row_level/numeric_checks/numeric_between_check.py
@@ -8,7 +8,7 @@ from sparkdq.checks.base_classes.base_comparison_check import BaseComparisonChec
 from sparkdq.core.base_config import BaseRowCheckConfig
 from sparkdq.core.severity import Severity
 from sparkdq.exceptions import InvalidCheckConfigurationError
-from sparkdq.factory.check_config_registry import register_check_config
+from sparkdq.plugin.check_config_registry import register_check_config
 
 
 class NumericBetweenCheck(BaseComparisonCheck):
