+++ sparkdq/checks/row_level/numeric_checks/numeric_min_check.py
@@ -7,7 +7,7 @@ from pyspark.sql import Column
 from sparkdq.checks.base_classes.base_comparison_check import BaseComparisonCheck
 from sparkdq.core.base_config import BaseRowCheckConfig
 from sparkdq.core.severity import Severity
-from sparkdq.factory.check_config_registry import register_check_config
+from sparkdq.plugin.check_config_registry import register_check_config
 
 
 class NumericMinCheck(BaseComparisonCheck):
